Adding property for ocr max distance. Refactoring packages. Optimize code.
