+++ src/main/resources/application.properties
@@ -1,4 +1,5 @@
 TESSDATA_PREFIX=tessdata
-MAX_DISTANCE=20
+MAX_DISTANCE=5
+MAX_DISTANCE_OCR=15
 spring.servlet.multipart.max-file-size=10MB
 spring.servlet.multipart.max-request-size=10MB
